+++ src/main/java/com/onlinebookstore/dto/user/UserLoginRequestDto.java
@@ -1,17 +1,17 @@
 package com.onlinebookstore.dto.user;
 
 import com.onlinebookstore.validation.Email;
-import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 
 @Data
 public class UserLoginRequestDto {
-    @NotEmpty
+    @NotNull
     @Size(min = 4, max = 50)
     @Email
     private String email;
-    @NotEmpty
+    @NotNull
     @Size(min = 6, max = 100)
     private String password;
 }
