+++ src/main/java/com/onlinebookstore/dto/user/UserLoginResponseDto.java
@@ -8,5 +8,3 @@ import lombok.RequiredArgsConstructor;
 public class UserLoginResponseDto {
     private final String token;
 }
-//public record UserLoginResponseDto(String token) {
-//}
