+++ src/main/java/com/onlinebookstore/dto/user/UserRegistrationRequestDto.java
@@ -2,7 +2,6 @@ package com.onlinebookstore.dto.user;
 
 import com.onlinebookstore.validation.Email;
 import com.onlinebookstore.validation.FieldMatch;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
@@ -15,16 +14,13 @@ import lombok.Data;
 )
 public class UserRegistrationRequestDto {
     @NotNull
-    @NotBlank
     @Size(min = 4, max = 50)
     @Email
     private String email;
     @NotNull
-    @NotBlank
     @Size(min = 6, max = 100)
     private String password;
     @NotNull
-    @NotBlank
     @Size(min = 6, max = 100)
     private String repeatPassword;
     @NotNull
