+++ src/main/java/com/onlinebookstore/repository/BookRepository.java
@@ -1,4 +1,4 @@
-package com.onlinebookstore.repository.book;
+package com.onlinebookstore.repository;
 
 import com.onlinebookstore.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
