+++ src/main/java/com/onlinebookstore/repository/TitleSpecificationProvider.java
@@ -1,7 +1,6 @@
-package com.onlinebookstore.repository.book.spec;
+package com.onlinebookstore.repository;
 
 import com.onlinebookstore.model.Book;
-import com.onlinebookstore.repository.SpecificationProvider;
 import jakarta.persistence.criteria.Predicate;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
