+++ src/main/java/com/onlinebookstore/service/impl/BookServiceImpl.java
@@ -6,8 +6,8 @@ import com.onlinebookstore.dto.book.CreateBookRequestDto;
 import com.onlinebookstore.exception.EntityNotFoundException;
 import com.onlinebookstore.mapper.BookMapper;
 import com.onlinebookstore.model.Book;
-import com.onlinebookstore.repository.book.BookRepository;
-import com.onlinebookstore.repository.book.BookSpecificationBuilder;
+import com.onlinebookstore.repository.BookRepository;
+import com.onlinebookstore.repository.spec.BookSpecificationBuilder;
 import com.onlinebookstore.service.BookService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
