+++ src/main/java/com/onlinebookstore/service/impl/UserServiceImpl.java
@@ -6,8 +6,8 @@ import com.onlinebookstore.exception.RegistrationException;
 import com.onlinebookstore.mapper.UserMapper;
 import com.onlinebookstore.model.Role;
 import com.onlinebookstore.model.User;
-import com.onlinebookstore.repository.role.RoleRepository;
-import com.onlinebookstore.repository.user.UserRepository;
+import com.onlinebookstore.repository.RoleRepository;
+import com.onlinebookstore.repository.UserRepository;
 import com.onlinebookstore.service.UserService;
 import java.util.HashSet;
 import java.util.Set;
